refactor: 도메인 계층 각 역할별 폴더 분리
