+++ src/main/java/domain/Judgement.java
@@ -1,5 +1,8 @@
 package domain;
 
+import domain.participant.Dealer;
+import domain.participant.Player;
+import domain.participant.Players;
 import java.util.LinkedHashMap;
 import java.util.Map;
