+++ src/main/java/domain/participant/Dealer.java
@@ -1,5 +1,7 @@
-package domain;
+package domain.participant;
 
+import domain.card.Card;
+import domain.card.Deck;
 import java.util.List;
 
 public class Dealer extends Participant {
