+++ src/main/java/domain/participant/Players.java
@@ -1,6 +1,5 @@
-package domain;
+package domain.participant;
 
-import java.util.HashSet;
 import java.util.List;
 
 public class Players {
