+++ src/test/java/domain/JudgementTest.java
@@ -1,5 +1,12 @@
 package domain;
 
+import domain.card.Card;
+import domain.card.CardNumber;
+import domain.card.CardSuit;
+import domain.card.Deck;
+import domain.participant.Dealer;
+import domain.participant.Player;
+import domain.participant.Players;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
