+++ src/test/java/domain/participant/PlayerTest.java
@@ -1,5 +1,9 @@
-package domain;
+package domain.participant;
 
+import domain.card.Card;
+import domain.card.CardNumber;
+import domain.card.CardSuit;
+import domain.card.Deck;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
